# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        },
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/simcards": {
            "get": {
                "tags": [
                    "Simcards"
                ],
                "summary": "Get a list of simcards",
                "operationId": "51682b5ab63673763249cd913d5c06d0",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Simcards"
                ],
                "summary": "Store a new simcard",
                "operationId": "34a4c88b45492b43aa210041c2c9bcb7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Simcard Name"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Simcard created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "number": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/simcards/{id}": {
            "get": {
                "tags": [
                    "Simcards"
                ],
                "summary": "Get a simcard",
                "operationId": "0a3cfc4d31811a2957dbf2e8111117ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Simcard ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "number": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Simcards"
                ],
                "summary": "Update a simcard",
                "operationId": "57d53c09992c37242bc4762c4ef4d8a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Simcard ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Simcard Name"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "0987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Simcard updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "number": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Simcards"
                ],
                "summary": "Delete a simcard",
                "operationId": "f5c774a41188b4245d17127a461030d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Simcard ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Simcard deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Simcard deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}